Feat: 타임라인 타입 피드 - 총 낚시 시간 표시

+++ app/src/main/java/com/android04/godfisherman/common/BindingAdapter.kt
@@ -108,7 +108,7 @@ object BindingAdapter {
 
     @JvmStatic
     @BindingAdapter("setTimeToMinute")
-    fun setTimeWithDouble(view: TextView, millisecond: Double) {
+    fun setTimeWithDouble(view: TextView, millisecond: Int) {
         view.text = millisecond.toTimeHourMinute()
     }
